Changed Shop interface and stuff
New buttons and camera border added input for which button is selected, also made ammo amount higher and based on the gun, the banners partially fixed changed ints to floats and it seems to "work" will need to do more to the saving of stats though

+++ Assets/Script/Player/loadStats.cs
@@ -42,10 +42,10 @@ public class loadStats : MonoBehaviour
             banner.hardSetKillCount("Poison", PlayerPrefs.GetInt("poisonKills"));
             banner.hardSetKillCount("Lightning", PlayerPrefs.GetInt("lightningKills"));
 
-            banner.setBannerAmount("Fire", PlayerPrefs.GetInt("fireBanner"));
-            banner.setBannerAmount("Water", PlayerPrefs.GetInt("waterBanner"));
-            banner.setBannerAmount("Poison", PlayerPrefs.GetInt("poisonBanner"));
-            banner.setBannerAmount("Lightning", PlayerPrefs.GetInt("lightningBanner"));
+            banner.setBannerAmount("Fire", PlayerPrefs.GetFloat("fireBanner"));
+            banner.setBannerAmount("Water", PlayerPrefs.GetFloat("waterBanner"));
+            banner.setBannerAmount("Poison", PlayerPrefs.GetFloat("poisonBanner"));
+            banner.setBannerAmount("Lightning", PlayerPrefs.GetFloat("lightningBanner"));
 
             Auto.hardSetCurrAmmo(PlayerPrefs.GetInt("AutoloadedAmmo"));
             Auto.hardSetUnloadedAmmo(PlayerPrefs.GetInt("AutostoredAmmo"));
